Made asteroids rotate and explode when hit by a players bullet
updated bullets to rotate as they float across screen. they also now take damage and explode when a bullet is detected in their hitbox. I think the hitbox needs to be a little refined because the bullet doesnt always line up to make the asteroid explode.
Needed to add position getters to Bullet